requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
dnspython==2.5.0
email-validator==2.1.0.post1
idna==3.6
psycopg2-binary==2.9.9
pytz==2023.4
sqlparse==0.4.4
tzdata==2023.4
